// src/common/enums/institutional-operational-action.enum.ts
export enum InstitutionalOperationalAction {
  INSTITUTIONAL_SUBJECT_CREATED = 'INSTITUTIONAL_SUBJECT_CREATED',
  PLANNING_VALIDATE_INITIAL = 'PLANNING_VALIDATE_INITIAL',
  PLANNING_RETURN_INITIAL = 'PLANNING_RETURN_INITIAL',
  FACTORY_START_PRODUCTION = 'FACTORY_START_PRODUCTION',
  FACTORY_DELIVER_CONTENT = 'FACTORY_DELIVER_CONTENT',
  PLANNING_VALIDATE_PRODUCTION = 'PLANNING_VALIDATE_PRODUCTION',
  PLANNING_RETURN_PRODUCTION = 'PLANNING_RETURN_PRODUCTION',
  LMS_START_UPLOAD = 'LMS_START_UPLOAD',
  LMS_CONFIRM_UPLOAD = 'LMS_CONFIRM_UPLOAD',
  PLANNING_VALIDATE_LMS = 'PLANNING_VALIDATE_LMS',
  PLANNING_RETURN_LMS = 'PLANNING_RETURN_LMS',
  PRODUCT_START_ACADEMIC_REVIEW = 'PRODUCT_START_ACADEMIC_REVIEW',
  PRODUCT_REQUEST_CHANGES = 'PRODUCT_REQUEST_CHANGES',
  PRODUCT_APPROVE_ACADEMIC = 'PRODUCT_APPROVE_ACADEMIC',
  PLANNING_FINALIZE = 'PLANNING_FINALIZE',
  PRODUCT_RESUBMIT_REQUEST = 'PRODUCT_RESUBMIT_REQUEST',
}