// src/common/enums/project-institutional-action.enum.ts
export enum ProjectInstitutionalAction {
  PRODUCT_REGISTER_RADICATION = 'PRODUCT_REGISTER_RADICATION',
  PRODUCT_RESUBMIT_RADICATION = 'PRODUCT_RESUBMIT_RADICATION',
  PLANNING_VALIDATE_RADICATION = 'PLANNING_VALIDATE_RADICATION',
  PLANNING_RETURN_RADICATION = 'PLANNING_RETURN_RADICATION',
  AUTO_READY_FOR_RADICATION = 'AUTO_READY_FOR_RADICATION',
}